StickBall2D - Experiment 2
Same as Experiment 1, with the introduction of a random initial angle.
This helps the neural network learn how to keep the stick balanced with
a more general range of angles.
In order for the stick to be balanced in situations with great angles,
the handle needs to have a higher velocity. Now the velocity can have
values higher than 1.0 up to an indicated maximum value.